+++ app/src/organisms/Devices/RobotSettings/__tests__/RobotSettingsAdvnaced.test.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+
+// test cases
+// 1 check text
+// 2 check link
+// 3 check button
+// 4 check toggle button
+// - MockReturnValue (true) and (false)
